PR feedback  - moved classes around and made the IDL spec support call
